upgraded validation for shop entities

+++ shop-service/src/main/java/com/vlados/webshop/shopservice/domain/dto/item/ItemRequestDto.java
@@ -11,7 +11,7 @@ public record ItemRequestDto(
         @NotBlank
         String categoryName,
         @Min(1)
-        int quantity,
+        long quantity,
         @HexColor
         String color,
         String description) {
